extracted download wrapper
