+++ src/widgets/SequencingIntensityPlot.tsx
@@ -7,10 +7,7 @@ import * as zod from 'zod';
 import { AsyncZodQueryEncoder } from '../helpers/query-encoder';
 import TimeIntensityChart, { TimeIntensityEntry } from '../charts/TimeIntensityChart';
 import Loader from '../components/Loader';
-import { exportComponentAsJPEG } from 'react-component-export-image';
-import { FaCloudDownloadAlt } from 'react-icons/fa';
-import ReactTooltip from 'react-tooltip';
-import styled, {keyframes} from 'styled-components';
+import DownloadWrapper from '../charts/DownloadWrapper';
 
 interface Props {
   country: Country;
@@ -72,52 +69,9 @@ export const SequencingIntensityPlot = ({ country }: Props) => {
   return data === undefined || isLoading ? (
     <Loader />
   ) : (
-    <ImageDownloadWrapper>
+    <DownloadWrapper>
       <TimeIntensityChart data={processData(data)} onClickHandler={(e: unknown) => true} />
-    </ImageDownloadWrapper>
-  );
-};
-
-const DownloadButton = styled(FaCloudDownloadAlt)`
-  position: absolute;
-  top: 0px;
-  right: 8px;
-  padding: 8px 8px 8px 8px;
-  &:hover {
-    cursor: pointer;
-    transform: translate(0%, -10%);
-    transition: 0.2s ease-out;
-  }
-`;
-
-const Wrapper = styled.div`
-  position: relative;
-`;
-
-//Adds button to download wrapper component as an image
-export const ImageDownloadWrapper = ({ fileName = 'plot.jpg', ...props }) => {
-  const componentRef = useRef(null);
-
-  const exportOptions = {
-    fileName: fileName,
-  };
-
-  return (
-    <>
-      <Wrapper>
-        <DownloadButton
-          data-for={fileName}
-          data-tip='Download this chart as JPEG image'
-          size='2.5em'
-          color='#95a5a6'
-          onClick={() => exportComponentAsJPEG(componentRef, exportOptions)}
-        />
-        <div id='image-download-container' ref={componentRef}>
-          {props.children}
-        </div>
-      </Wrapper>
-      <ReactTooltip id={fileName} delayShow={1000}/>
-    </>
+    </DownloadWrapper>
   );
 };
